part two of the single crawl page UI for owners and non owners
